pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "psy2"
version = "0.1.0"
description = "Psychology AI RAG Patient Simulator"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "sqlalchemy==2.0.34",
    "asyncpg==0.29.0",
    "alembic==1.13.2",
    "psycopg[binary]==3.2.1",
    "pgvector==0.2.5",
    "redis==5.0.8",
    "httpx==0.27.2",
    "sentence-transformers==3.0.1",
    "loguru==0.7.2",
    "prometheus-client==0.20.0",
    "opentelemetry-sdk==1.27.0",
    "opentelemetry-exporter-otlp==1.27.0",
    "opentelemetry-instrumentation-fastapi==0.48b0",
    "opentelemetry-instrumentation-httpx==0.48b0",
    "pytest==8.3.2",
    "hypothesis==6.112.3",
    "greenlet==3.2.4",
    "click==8.1.7",
    "tenacity==8.5.0",
    "json5==0.9.25",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E","F","I"]

[tool.ruff.lint.isort]
known-first-party = ["app","tests"]

[tool.pytest.ini_options]
asyncio_mode = "strict"


[tool.poetry.group.dev.dependencies]
fakeredis = "==2.23.2"

